third_party/blink/web_tests/plugins/tabindex-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
Test for the tabIndex of embedded plugins. Plugins may be focusable so their tabIndex should be 0 by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "embedElem"; pluginElement.tabIndex is -1
PASS "objectElem"; pluginElement.tabIndex is 1
PASS "embedElemWithFallbackContents"; pluginElement.tabIndex is -1
PASS "objectElemWithFallbackContents"; pluginElement.tabIndex is 0
PASS "noPluginEmbedElem"; pluginElement.tabIndex is 1
PASS "noPluginObjectElem"; pluginElement.tabIndex is 0
PASS "noPluginEmbedElemWithFallbackContents"; pluginElement.tabIndex is 0
PASS "noPluginObjectElemWithFallbackContents"; pluginElement.tabIndex is 0
PASS "noPluginEmbedElemWithDisplayNone"; pluginElement.tabIndex is -1
PASS "noPluginObjectElemWithDisplayNone"; pluginElement.tabIndex is 0
PASS "noPluginEmbedElemWithContenteditable"; pluginElement.tabIndex is -1
PASS "noPluginObjectElemWithContenteditable"; pluginElement.tabIndex is -1
PASS successfullyParsed is true

TEST COMPLETE

  Fallback contents.    Fallback contents. Fallback contents.  
